pyproject.toml
[tool.poetry]
name = "commit-message-gen"
version = "0.1.0"
description = ""
readme = "README.md"
authors = ["shuhei fujita <syuheifujita0117@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "*"
openai = "0.28.1"
pre-commit = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
